+++ docs/additional_information/changelog.rst
@@ -11,6 +11,9 @@ Added
 - Added option to automatically compute the window size for various anomaly detectors
   using Fourier Transform, Autocorrelation, Multi-Window Finder, or Summary Statistics Subsequence.
 - Implemented ``KNearestNeighbor`` anomaly detector.
+- Implemented point-adjusted precision, recall and FBeta.
+- Implemented ``BestThresholdMetric``, a ``ProbaMetric`` which computes the best value for
+  a ``BinaryMetric`` over all thresholds.
 
 Changed
 ^^^^^^^